--- src/components/assistant/HowItWorks.tsx
@@ -1,262 +1,0 @@
-import { Stack, Typography, useTheme } from "@mui/material";
-import Box from "@mui/material/Box";
-import React, { forwardRef, useCallback, useEffect, useImperativeHandle, useMemo, useRef, useState } from "react";
-
-import { useWindowSize } from "@/hooks/useWindowSize";
-
-import { gray03 } from "../../pages/assistant";
-// import { gray03 } from "../../pages";
-import { RiveComponentMemoized } from "../home/components/temporals/RiveComponentExtended";
-
-// import { useWindowSize } from "../../hooks/useWindowSize";
-// import Collapse from "@mui/material/Collapse";
-// import { CardActionArea } from "@mui/material";
-// import Typography from "../components/Typography";
-// import CustomTypography from "../components/Typography";
-// import { sectionsOrder } from "../sectionsOrder";
-// import { sectionsOrder } from "../sectionsOrder";
-// // import sectionsOrder from "./sectionsOrder";
-
-// const sectionIdx = sectionsOrder.findIndex(sect => sect.id === "HowItWorksSection");
-type HowItWorksProps = { artboards: any[] };
-
-const HowItWorks = ({ artboards }: HowItWorksProps, ref: any) => {
-  const [canvasDimension, setCanvasDimension] = useState({ width: 0, height: 0 });
-  const ref1 = useRef<any | null>(null);
-  const ref2 = useRef<any | null>(null);
-  const ref3 = useRef<any | null>(null);
-  // const ref4 = useRef<any | null>(null);
-  // const ref5 = useRef<any | null>(null);
-
-  const refs = useMemo(() => [ref1, ref2, ref3 /* , ref4, ref5 */], []);
-
-  useImperativeHandle(ref, () => ({
-    getHeight1: () => (ref1?.current ? ref1.current.clientHeight : 0),
-    getHeight2: () => (ref2?.current ? ref2.current.clientHeight : 0),
-    getHeight3: () => (ref3?.current ? ref3.current.clientHeight : 0),
-    // getHeight4: () => (ref4?.current ? ref4.current.clientHeight : 0),
-    // getHeight5: () => (ref5?.current ? ref5.current.clientHeight : 0),
-  }));
-
-  const { width } = useWindowSize({ initialHeight: 1000, initialWidth: 0 });
-  const theme = useTheme();
-
-  const getGrayColorText = useCallback(
-    () => (theme.palette.mode === "dark" ? gray03 : theme.palette.common.darkBackground2),
-    [theme.palette.common.darkBackground2, theme.palette.mode]
-  );
-
-  useEffect(() => {
-    let newWidth = width / 2;
-    if (width >= 1536) newWidth = 850;
-    else if (width >= 1200) newWidth = 750;
-    else if (width >= 900) newWidth = 700;
-    else if (width >= 600) newWidth = 540;
-    else if (width >= 375) newWidth = 370;
-    else if (width >= 0) newWidth = width - 20;
-
-    const newHeight = getHeight(newWidth);
-    setCanvasDimension({ width: newWidth, height: newHeight });
-  }, [width]);
-
-  // const boxLarge = useMemo(() => {
-  //   const offset = width < 600 ? 32 : width < 900 ? 70 : 100;
-  //   if (height < width) return height - offset;
-  //   return width - offset;
-  // }, [height, width]);
-
-  // const topCenteredPosition = height / 2 - boxLarge / 2 + 35;
-
-  // const getHeightSection = () => props.artboards.reduce((a: number, c: any) => a + c.getHeight(height), 0);
-
-  // const processedArtboard = useMemo(
-  //   () =>
-  //     props.artboards.reduce((acu: any[], cur: any) => {
-  //       const newHeight = cur.getHeight(height);
-  //       return [
-  //         ...acu,
-  //         { ...cur, top: acu.length ? acu[acu.length - 1].top + acu[acu.length - 1].height : 0, height: newHeight },
-  //       ];
-  //     }, []),
-  //   [props.artboards, height]
-  // );
-
-  // return (
-  //   <Box
-  //     // id="HowItWorksSection"
-  //     component="section"
-  //     sx={{
-  //       height: getHeightSection(),
-  //       position: "relative",
-  //       display: "flex",
-  //       flexDirection: "column",
-  //       alignItems: "center",
-  //     }}
-  //   >
-  //     <Box
-  //       key={"artboard-1"}
-  //       sx={{
-  //         position: "absolute",
-  //         top: 0,
-  //         width: "100%",
-  //         height: height - 70,
-  //         // borderRight: `dashed 6px #ff28c9`,
-  //         color: "white",
-  //       }}
-  //     >
-  //     </Box>
-  //     <CustomTypography
-  //       component={"h2"}
-  //       variant="h1"
-  //       marked="center"
-  //       align="center"
-  //       sx={{ mb: 7, fontWeight: 700, position: "absolute", top: height - 30 }}
-  //     >
-  //       {sectionsOrder[1].title}
-  //     </CustomTypography>
-  //     {processedArtboard.map((artboard: any, idx: number) => (
-  //       <Box
-  //         key={artboard.name}
-  //         sx={{
-  //           position: "absolute",
-  //           top: artboard.top,
-  //           width: "100%",
-  //           height: artboard.height,
-  //           // borderRight: `dashed 6px ${artboard.color}`,
-  //           color: "white",
-  //         }}
-  //       >
-  //         {idx > 0 && (
-  //           <Typography
-  //             variant="h5"
-  //             component="h3"
-  //             sx={{
-  //               mt: "100px",
-  //               ml: "10px",
-  //               position: "sticky",
-  //               top: "100px",
-  //               // color: "white",
-  //               textTransform: "none",
-  //             }}
-  //           >
-  //             {artboard.name}
-  //           </Typography>
-  //         )}
-  //       </Box>
-  //     ))}
-
-  //     <Box sx={{ position: "absolute", bottom: "20px", zIndex: 11 }}>{props.animationOptions}</Box>
-
-  //     <div
-  //       style={{
-  //         position: "sticky",
-  //         top: topCenteredPosition,
-  //         width: boxLarge,
-  //         height: boxLarge,
-  //         display: "flex",
-  //         flexDirection: "column",
-  //         zIndex: 10,
-  //         // border: "solid 2px pink"
-  //       }}
-  //     >
-  //       {props.children}
-  //     </div>
-  //   </Box>
-  // );
-  const AnimationSections = useMemo(() => {
-    return artboards.map((artboard: any, idx: number, src: any[]) => (
-      <Stack
-        ref={refs[idx]}
-        key={artboard.name}
-        direction={"column"}
-        spacing={"20px"}
-        alignItems={width < 900 ? "center" : "stretch"}
-        sx={{ position: "relative", minHeight: "500px" /* , border: `2px dashed red` */ }}
-      >
-        <Typography
-          gutterBottom
-          variant="h3"
-          component="h3"
-          sx={{ fontSize: "32px", textAlign: "center", fontWeight: 600 }}
-        >
-          {artboard.name}
-        </Typography>
-        <Box sx={{ position: "relative", alignSelf: "center" }}>
-          <Box
-            sx={{
-              position: "relative",
-
-              display: "flex",
-              alignItems: "center",
-            }}
-          >
-            <Box
-              sx={{
-                width: `${canvasDimension.width}px`,
-                height: idx === 2 ? `${canvasDimension.height - 80}px` : `${canvasDimension.height}px`,
-              }}
-            >
-              <RiveComponentMemoized
-                src={artboard.src}
-                artboard={artboard.artoboard}
-                animations={["Timeline 1", theme.palette.mode]}
-                autoplay={true}
-              />
-            </Box>
-          </Box>
-        </Box>
-
-        <Box
-          sx={{
-            p: "10px",
-            display: "flex",
-            flexDirection: "column",
-            justifyContent: "center",
-            "& > *:not(:last-child)": {
-              mb: "12px",
-            },
-            pb: idx < src.length - 1 ? "50px" : "0px",
-          }}
-        >
-          {artboard.description.split("\n").map((paragraph: string, idx: number) => (
-            <Typography
-              key={idx}
-              variant="body2"
-              sx={{
-                textAlign: "left",
-                color: getGrayColorText(),
-                fontSize: "16px",
-              }}
-            >
-              {paragraph}
-            </Typography>
-          ))}
-        </Box>
-      </Stack>
-    ));
-  }, [artboards, refs, width, canvasDimension.width, canvasDimension.height, theme.palette.mode, getGrayColorText]);
-
-  return (
-    <Stack
-      component="section"
-      direction={"column"}
-      alignItems={"center"}
-      spacing={width < 900 ? "0px" : "20px"}
-      sx={{
-        // position: "relative",
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center",
-      }}
-    >
-      {AnimationSections}
-    </Stack>
-  );
-};
-
-// const getHeight = (width: number) => (300 * width) / 500;
-const getHeight = (width: number) => width;
-
-const HowItWorksFordwarded = forwardRef(HowItWorks);
-
-export default HowItWorksFordwarded;
