--- src/components/assistant/sectionsOrder.ts
@@ -1,7 +1,0 @@
-export const sectionsOrder = [
-  { id: "Landing", label: "Hero", title: "Hero" },
-  { id: "How1CademyAssistantWorks", label: "How", title: "How 1Cademy Assistant works?" },
-  { id: "Why1CademyAssistant", label: "Why", title: "Why 1Cademy Assistant?" },
-  { id: "WhichSection", label: "Which", title: "Which systems?" },
-  { id: "WhoIsBehind1CademyAssistant", label: "Who", title: "Who's Behind 1Cademy Assistant?" },
-];
