+++ src/components/home/sections/constants/valuesItems.ts
